requirements.txt
streamlit==1.35.0
opencv-python==4.9.0.80
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
deep-translator==1.11.4
google-generativeai==0.4.1
pydantic==2.7.1
httpx==0.27.0


